+++ src/colloids.c
@@ -7,7 +7,7 @@
  *  Edinburgh Soft Matter and Statistical Physics Group and
  *  Edinburgh Parallel Computing Centre
  *
- *  (c) 2010-2021 The University of Edinburgh
+ *  (c) 2010-2023 The University of Edinburgh
  *
  *  Contributing authors:
  *  Kevin Stratford (kevin@epcc.ed.ac.uk)
@@ -1288,40 +1288,6 @@ __host__ int colloids_info_ahmax(colloids_info_t * cinfo, double * ahmax) {
   for (; pc; pc = pc->next) ahmax_local = dmax(ahmax_local, pc->s.ah);
 
   MPI_Allreduce(&ahmax_local, ahmax, 1, MPI_DOUBLE, MPI_MAX, comm);
-
-  return 0;
-}
-
-/*****************************************************************************
- *
- *  colloids_info_count_local
- *
- *  Return number of local colloids of given type.
- *
- *****************************************************************************/
-
-__host__ int colloids_info_count_local(colloids_info_t * cinfo,
-				       colloid_type_enum_t it,
-				       int * count) {
-  int nlocal = 0;
-  int ic, jc, kc;
-  colloid_t * pc = NULL;
-
-  assert(cinfo);
-
-  for (ic = 1; ic <= cinfo->ncell[X]; ic++) {
-    for (jc = 1; jc <= cinfo->ncell[Y]; jc++) {
-      for (kc = 1; kc <= cinfo->ncell[Z]; kc++) {
-
-	colloids_info_cell_list_head(cinfo, ic, jc, kc, &pc);
-	for (; pc; pc = pc->next) {
-	  if (pc->s.type == it) nlocal += 1;
-	}
-      }
-    }
-  }
-
-  *count = nlocal;
 
   return 0;
 }