+++ src/subgrid.c
@@ -126,7 +126,7 @@ int subgrid_force_from_particles(colloids_info_t * cinfo, hydro_t * hydro,
 
 	for ( ; p_colloid; p_colloid = p_colloid->next) {
 
-          if (p_colloid->s.type != COLLOID_TYPE_SUBGRID) continue;
+	  if (p_colloid->s.bc != COLLOID_BC_SUBGRID) continue;
 
 	  /* Need to translate the colloid position to "local"
 	   * coordinates, so that the correct range of lattice
@@ -252,7 +252,7 @@ int subgrid_update(colloids_info_t * cinfo, hydro_t * hydro, int noise_flag) {
 
 	for ( ; p_colloid; p_colloid = p_colloid->next) {
 
-          if (p_colloid->s.type != COLLOID_TYPE_SUBGRID) continue;
+	  if (p_colloid->s.bc != COLLOID_BC_SUBGRID) continue;
 
 	  drag = reta*(1.0/p_colloid->s.ah - 1.0/p_colloid->s.al);
 
@@ -336,7 +336,7 @@ static int subgrid_interpolation(colloids_info_t * cinfo, hydro_t * hydro) {
 
 	for ( ; p_colloid; p_colloid = p_colloid->next) {
 
-          if (p_colloid->s.type != COLLOID_TYPE_SUBGRID) continue;
+	  if (p_colloid->s.bc != COLLOID_BC_SUBGRID) continue;
 
 	  p_colloid->fsub[X] = 0.0;
 	  p_colloid->fsub[Y] = 0.0;
@@ -356,7 +356,7 @@ static int subgrid_interpolation(colloids_info_t * cinfo, hydro_t * hydro) {
 
 	for ( ; p_colloid; p_colloid = p_colloid->next) {
 
-          if (p_colloid->s.type != COLLOID_TYPE_SUBGRID) continue;
+	  if (p_colloid->s.bc != COLLOID_BC_SUBGRID) continue;
 
 	  /* Need to translate the colloid position to "local"
 	   * coordinates, so that the correct range of lattice
@@ -431,7 +431,7 @@ int subgrid_wall_lubrication(colloids_info_t * cinfo, wall_t * wall) {
   colloids_info_local_head(cinfo, &pc);
 
   for ( ; pc; pc = pc->nextlocal) {
-    if (pc->s.type != COLLOID_TYPE_SUBGRID) continue;
+    if (pc->s.bc != COLLOID_BC_SUBGRID) continue;
     wall_lubr_sphere(wall, pc->s.ah, pc->s.r, drag);
     pc->fex[X] += drag[X]*pc->s.v[X];
     pc->fex[Y] += drag[Y]*pc->s.v[Y];