+++ tests/unit/test_colloid.c
@@ -44,8 +44,8 @@ int test_colloid_suite(void) {
 
   colloid_state_t sref = {1, 2, 3, 4, 5, 6, 7, 8, 9,
 			  {10, 11}, 12, {13, 14, 15}, {16, 17, 18}, 19,
-			  {20, 21, 22, 23, 24,
-			   25, 26, 27, 28, 29, 30, 31, 32},
+			  20, 21, 22, 23, 24, 25,
+			   {26, 27, 28, 29, 30, 31, 32},
 			  1.0, 2.0,
 			  { 3.0,  4.0,  5.0},
 			  { 6.0,  7.0,  8.0},
@@ -68,7 +68,7 @@ int test_colloid_suite(void) {
   /* I assert that the colloid struct is 512 bytes. I.e., don't
    * change it without sorting out the padding. */
   test_assert(sizeof(colloid_state_t) == 512);
-  assert(NPAD_INT == 13);
+  assert(NPAD_INT ==  7);
   assert(NPAD_DBL ==  4);
 
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
@@ -179,8 +179,6 @@ void test_colloid_binary_io(colloid_state_t * sref, const char * filename) {
 
 int test_colloid_compare(colloid_state_t * s1, colloid_state_t * s2) {
 
-  int n;
-
   test_assert(s1->index    == s2->index);
   test_assert(s1->nbonds   == s2->nbonds);
   test_assert(s1->nangles  == s2->nangles);
@@ -190,7 +188,8 @@ int test_colloid_compare(colloid_state_t * s1, colloid_state_t * s2) {
   test_assert(s1->isfixeds == s2->isfixeds);
   test_assert(s1->type     == s2->type);
 
-  for (n = 0; n < NBOND_MAX; n++) {
+  assert(NBOND_MAX == 2);
+  for (int n = 0; n < NBOND_MAX; n++) {
     test_assert(s1->bond[n] == s2->bond[n]);
   }