+++ tests/unit/test_ewald.c
@@ -10,7 +10,7 @@
  *  Edinburgh Soft Matter and Statistical Physics Group and
  *  Edinburgh Parallel Computing Centre
  *
- *  (c) 2010-2022 The University of Edinburgh
+ *  (c) 2010-2023 The University of Edinburgh
  *
  *  Contributing authors:
  *  Kevin Stratford (kevin@epcc.ed.ac.uk)
@@ -112,6 +112,7 @@ int test_ewald_suite(void) {
   p_c1->s.s[X] = 0.0;
   p_c1->s.s[Y] = 0.0;
   p_c1->s.s[Z] = 1.0;
+  p_c1->s.magnetic = 1;
 
   /* Second colloid ... */
 
@@ -121,6 +122,7 @@ int test_ewald_suite(void) {
   p_c2->s.s[X] = 0.0;
   p_c2->s.s[Y] = 0.0;
   p_c2->s.s[Z] = -1.0;
+  p_c2->s.magnetic = 1;
   cs_minimum_distance(cs, r1, r2, r12);