+++ CHANGES.md
@@ -1,6 +1,15 @@
 
 ### Changes
 
+- "colloid_type" has been split into a number of more general
+  properties including: boundary condition (bc) either bbl or
+  subgrid; shale either "disk", "sphere", or "ellipsoid";
+  active (logical); magnetic (logical).
+  The default is: bc = bbl, shape = sphere, active = no, magnetic = no.
+  There are some conditions, e.g., "disk" expects a 2D D2Q9 model.
+  This affects the way colloid details are specified in the input.
+  See X
+
 version 0.20.0
 
 - Added coverage via https://about.codecov.io/